requirements.txt
# Core ML/DL Libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
adapters>=0.2.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# NLP Tools
nltk>=3.8.0
sentencepiece>=0.1.99
tokenizers>=0.13.0

# Data Augmentation
nlpaug>=1.1.11
textaugment>=1.3.4

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Logging and Monitoring
wandb>=0.15.0
tensorboard>=2.13.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.23.0
matplotlib>=3.7.0
seaborn>=0.12.0
